migrate funcitonAdapter to own folder
