+++ packages/data-mate/src/adapters/data-mate/index.ts
@@ -1,0 +1,3 @@
+export function dataMateFunctionAdapter() {
+    throw new Error('not implemented yet');
+}
